Rename `hotkeys_provider` module as `context`
